+++ app/src/main/java/com/teamwss/websoso/ui/keyword/KeywordActivity.kt
@@ -1,12 +1,12 @@
-package com.teamwss.websoso.ui.view.memowrite
+package com.teamwss.websoso.ui.keyword
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.teamwss.websoso.R
 
-class MemoWriteActivity : AppCompatActivity() {
+class KeywordActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_memo_write)
+        setContentView(R.layout.activity_keyword)
     }
 }
