+++ app/src/main/java/com/teamwss/websoso/ui/main/library/LibraryFragment.kt
@@ -1,4 +1,4 @@
-package com.teamwss.websoso.ui.view.main.mypage
+package com.teamwss.websoso.ui.main.library
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,13 +7,13 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.teamwss.websoso.R
 
-class MyPageFragment : Fragment() {
+class LibraryFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_my_page, container, false)
+        return inflater.inflate(R.layout.fragment_library, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -21,6 +21,6 @@ class MyPageFragment : Fragment() {
     }
 
     companion object {
-        fun newInstance() = MyPageFragment()
+        fun newInstance() = LibraryFragment()
     }
 }
