+++ app/src/main/java/com/teamwss/websoso/ui/postNovel/PostNovelActivity.kt
@@ -1,12 +1,12 @@
-package com.teamwss.websoso.ui.view.noveldetail
+package com.teamwss.websoso.ui.postNovel
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.teamwss.websoso.R
 
-class NovelDetailActivity : AppCompatActivity() {
+class PostNovelActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_novel_detail)
+        setContentView(R.layout.activity_post_novel)
     }
 }
