+++ app/src/main/res/layout/activity_keyword.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.view.keyword.KeywordActivity">
+    tools:context=".ui.keyword.KeywordActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
