+++ app/src/main/res/layout/activity_novel_detail.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.view.noveldetail.NovelDetailActivity">
+    tools:context=".ui.novelDetail.NovelDetailActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
