+++ app/src/main/res/layout/activity_post_novel.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.view.postnovel.PostNovelActivity">
+    tools:context=".ui.postNovel.PostNovelActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
