+++ app/src/main/res/layout/activity_search.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.view.search.SearchActivity">
+    tools:context=".ui.search.SearchActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
